chore: move drawer components to folder components
